+++ contracts/token/ERC721/extensions/ERC721Enumerable.sol
@@ -169,7 +169,7 @@ abstract contract ERC721Enumerable is ERC721, IERC721Enumerable {
     /**
      * See {ERC721-_increaseBalance}. We need that to account tokens that were minted in batch
      */
-    function _increaseBalance(address account, uint256 amount) internal virtual override {
+    function _increaseBalance(address account, uint128 amount) internal virtual override {
         if (amount > 0) {
             revert ERC721EnumerableForbiddenBatchMint();
         }
